+++ chainerx_cc/chainerx/native/native_device.h
@@ -107,8 +107,6 @@ public:
     void IsNan(const Array& x, const Array& out) override;
     void IsInf(const Array& x, const Array& out) override;
 
-    void Ceil(const Array& x, const Array& out) override;
-
     // conv.cc
 
     Array Conv(
